requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
html5tagger==1.3.0
httptools==0.5.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.2
psycopg2-binary==2.9.5
pyrsistent==0.19.3
requests==2.28.2
sanic==23.3.0
sanic-routing==22.8.0
tqdm==4.65.0
tracerite==1.1.0
ujson==5.7.0
urllib3==1.26.15
uvloop==0.17.0
websockets==10.4
yarl==1.8.2
